delete parcel cache
