--- nature-nomads/.parcel-cache/snapshot-39676f90f93af3b1.txt
@@ -1,2 +1,0 @@
-580034496
-1722909793345174000
