+++ nature-nomads/public/js/index.js
@@ -1,1 +1,10 @@
-console.log('Hello from parcel!');
+import { login } from './login';
+
+document.querySelector('.form').addEventListener('submit', (e) => {
+  e.preventDefault();
+
+  const email = document.getElementById('email').value;
+  const password = document.getElementById('password').value;
+
+  login(email, password);
+});
