+++ nature-nomads/public/js/login.js
@@ -1,6 +1,6 @@
 /* eslint-disable */
 
-const login = async (email, password) => {
+export const login = async (email, password) => {
   try {
     const res = await axios({
       method: 'POST',
@@ -19,12 +19,3 @@ const login = async (email, password) => {
     alert(e.response.data.message);
   }
 };
-
-document.querySelector('.form').addEventListener('submit', (e) => {
-  e.preventDefault();
-
-  const email = document.getElementById('email').value;
-  const password = document.getElementById('password').value;
-
-  login(email, password);
-});
