+++ nature-nomads/views/base.pug
@@ -22,7 +22,5 @@ html
         // FOOTER
         include components/_footer
 
-        // RUN SCRIPT TO LOAD MAP
+        // RUN SCRIPT TO LOAD ALL CODE & DEPENDENCIES
         script(src='/js/bundle.js')
-     
-
